+++ lib/src/relations/morph_many.dart
@@ -1,5 +1,6 @@
 import 'relation.dart';
 import '../core/model.dart';
+import '../core/query_builder.dart';
 
 /// Defines a polymorphic one-to-many relationship.
 ///
@@ -38,18 +39,21 @@ class MorphMany<R extends Model> extends Relation<R> {
   Future<void> match(
     List<Model> models,
     String relationName, {
-    List<String> nested = const [],
+    ScopeCallback? scope,
+    Map<String, ScopeCallback?> nested = const {},
   }) async {
     final ids = getKeys(models, parent.primaryKey);
 
-    final results =
-        (await creator({})
-                .newQuery()
-                .withRelations(nested)
-                .where('${name}_type', type)
-                .whereIn('${name}_id', ids)
-                .get())
-            .cast<R>();
+    final query = creator({}).newQuery();
+    query.withRelations(nested);
+    query.where('${name}_type', type);
+    query.whereIn('${name}_id', ids);
+
+    if (scope != null) {
+      scope(query);
+    }
+
+    final results = (await query.get()).cast<R>();
 
     for (var model in models) {
       final myId = normKey(model.id);